Move location filter generation functions
